requirements/base.txt
# Django
Django==5.0.4
djangorestframework==3.15.1
django-filter==23.5
django-environ==0.11.2
djoser==2.2.2
djangorestframework-simplejwt==5.3.1
social-auth-app-django==5.4.0
drf-nested-routers==0.94.1

# Security
argon2-cffi==23.1.0
cryptography==42.0.7

# Payment
mercadopago==2.2.3

# Docs
drf-spectacular==0.27.2

# Redis
django-redis==5.4.0
redis==5.0.2

# Utilities
pillow==10.3.0
django-simple-history==3.7.0
